04_ai_engine_service/requirements-railway.txt
# StudyAI AI Engine - Railway Deployment Requirements

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI Integration
openai==1.3.7
tiktoken==0.5.1

# Authentication & Security  
PyJWT==2.8.0
cryptography==41.0.7

# Session Storage
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Educational Processing (lightweight)
numpy==1.25.2